bugfix: fix conditionals to test emptiness
when you test a variable, and its value is a string, recent `ansible`
throws a cryptic error.

```
The conditional check 'test' failed. The error was: expected token 'end
of statement block', got '...'
```

where '...' is the string.

here is what the dev says:
https://github.com/ansible/ansible/issues/20392#issuecomment-280430379

> that actually happens on older versions of Ansible as well, so that
> case is not a regression, it's just a side-effect of the fact that we
> allow bare variables like this in conditionals. The correct solution is
> to use is `defined` in these kinds of situations where the variable may
> not contain a boolean value.

in short, when the variable is `true` or `false`, it works as expected.

```yaml

- name: foo
  ...
  when: true_or_false
```

but when the variable is not boolean, use `defined` and `!=''` to test
emptiness.