do NOT use `enabled: true` in handlers
in most cases, you set `enbled: true` in a service task in tasks. it
might look same when you set `enabled: true` in BOTH a service task and
a service handler, but it would bites you when you have `arguments` in a
service task.

`arguments` is not implemented in all platforms. it is in OpenBSD. when
both the service task and the handler has `enabled: true`, but only
one of them has `arguments`, then, `idempotency_test` will mysteriously
fail because the one with `arguments` modifies `rc.conf.local` and the
other as well.

this happens when:

* the platform is OpenBSD
* a service in task and a service in handler has `enabled: true`
* only one of them has `arguments`
* `arguments` is not identical to the default